[AIRFLOW-3277] Correctly observe DST transitions for cron (#4117)
`following_schedule` converts to naive time by using the
local time zone. In case of a DST transition, say 3AM -> 2AM
("summer time to winter time") we incorrectly re-applied
the timezone information which meant that a "CEST -> CEST"
could happen instead of a "CEST -> CET". This resulted
in infinite loops.